move ProgressUpdater interface to UIService
